+++ src/ld53/score.ts
@@ -9,7 +9,7 @@ import { createAABB, pointInAABB } from "../physics/aabb.js";
 import { PhysicsStateDef } from "../physics/nonintersection.js";
 import { PhysicsParentDef } from "../physics/transform.js";
 import { TimeDef } from "../time/time.js";
-import { WoodHealthDef } from "../wood/wood-builder.js";
+import { WoodHealthDef } from "../wood/wood-health.js";
 import { MapPaths } from "../levels/map-loader.js";
 import { LD52ShipDef } from "./ship.js";
 import { Phase } from "../ecs/sys-phase.js";
